refactor: Consolidate UI into single Dioxus app module (#89)
- Move Layout, Nav, ThemeSwitcher components to src/app/components/
- Update all pages to use shared Layout component with nav/footer
- Delete src/ui/ module (now fully replaced by src/app/)
- Update lib.rs to remove ui module export
- Update main.rs with inline flash_page, control_redirect, settings_redirect
- Update integration tests to use stub handlers for UI routes

This completes the migration from the dual UI system (src/ui + src/app)
to a single consolidated Dioxus fullstack app.
